Enable cell stratification file option to properly update arcs track filtering.
Issue was that arcs track updates were deep cloning the default rather than modifying the existing possibly already modified config
